dom/nodes/Node-appendChild-three-scripts.html

<!doctype html>
<meta charset=utf-8>
<title>Node.appendChild: inserting three scripts from a div</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-node-appendChild">
<link rel=help href="https://dom.spec.whatwg.org/#concept-node-insert">
<link rel=help href="https://github.com/whatwg/dom/issues/575">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
</head>
<body>
<div id="log"></div>
<script>
var s1 = document.createElement("script");
var s2 = document.createElement("script");
var s3 = document.createElement("script");

var s1Ran = false;
s1.textContent = `
    s3.appendChild(new Text("s3Ran = true;"));
    assert_true(s3Ran, "s3 did not run");
    s1Ran = true;
`;

var s2Ran = false;
s2.textContent = `
    assert_true(s1Ran, "s1 did not run");
    s2Ran = true;
`;

var s3Ran = false;
s3.textContent = `
    assert_false(s1Ran, "s1 already ran");
`;

var div = document.createElement("div");
div.appendChild(s1);
div.appendChild(s2);
div.appendChild(s3);

assert_false(s1Ran, "s1 ran");
document.body.appendChild(div);
assert_true(s2Ran, "s2 did not run");
done();
</script>
